pyproject.toml
[project]
name = "cinebaianos-api"
version = "1.0.0"
description = ""
authors = [ { name = "Bruno Reolon", email = "bruno.reolonn@email.com" } ]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "oauth2client==4.1.3",
    "requests==2.32.4",
    "python-dotenv==1.1.1",
    "tmdbv3api==1.9.0",
    "google-auth==2.40.3",
    "gspread==6.2.1",
    "gspread-formatting==1.2.1",
    "flask==3.1.1"
]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]